d3bfed/requirements.txt
Flask==2.1.2
pymongo==4.3.3
python-dotenv~=0.21.0